requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.0
oauthlib==3.1.0
Pillow==7.1.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.17
tweepy==3.8.0
tweet-preprocessor==0.6.0
urllib3==1.25.9
vaderSentiment==3.3.2
Werkzeug==1.0.1
wordcloud==1.7.0
